requirements.txt
prophet
streamlit
anthropic
pandas
plotly
python-dotenv
numpy
